+++ src/main/java/net/jonathangiles/tools/teenyhttpd/json/TeenyJson.java
@@ -1,6 +1,9 @@
 package net.jonathangiles.tools.teenyhttpd.json;
 
 
+import net.jonathangiles.tools.teenyhttpd.implementation.ParameterType;
+import net.jonathangiles.tools.teenyhttpd.implementation.ReflectionUtils;
+
 import java.io.BufferedOutputStream;
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
